@@ -24,7 +24,7 @@ const routes = [
 ];
 
 const router = createRouter({
-	history: createWebHashHistory(process.env.BASE_URL),
+	history: createWebHashHistory(import.meta.env.BASE_URL),
 	routes,
 });
 
